+++ index.d.ts
@@ -11,14 +11,18 @@ declare module 'react-native-ffmpeg' {
         static AV_LOG_VERBOSE: number;
         static AV_LOG_DEBUG: number;
         static AV_LOG_TRACE: number;
+
+        static logLevelToString(number): string;
     }
 
-    export interface LogMessage {
+    export interface Log {
+        executionId: number;
         level: LogLevel;
-        log: string;
+        message: string;
     }
 
     export interface Statistics {
+        executionId: number;
         videoFrameNumber: number;
         videoFps: number;
         videoQuality: number;
@@ -28,71 +32,94 @@ declare module 'react-native-ffmpeg' {
         speed: number;
     }
 
-    export interface StreamInformation {
-        index?: number;
-        type?: string;
-        codec?: string;
-        fullCodec?: string;
-        format?: string;
-        fullFormat?: string;
-        width?: number;
-        height?: number;
-        bitrate?: number;
-        sampleRate?: number;
-        sampleFormat?: string;
-        channelLayout?: string;
-        sampleAspectRatio?: string;
-        displayAspectRatio?: string;
-        averageFrameRate?: string;
-        realFrameRate?: string;
-        timeBase?: string;
-        codecTimeBase?: string;
-        metadata?: [string, string];
-        sidedata?: [string, string];
+    export interface Execution {
+        executionId: number;
+        startTime: number;
+        command: string;
     }
 
-    export interface MediaInformation {
-        format?: string;
-        path?: string
-        duration?: number;
-        startTime?: number;
-        bitrate?: number;
-        metadata?: [string, string];
-        streams?: StreamInformation[];
-        rawInformation?: string;
+    export interface CompletedExecution {
+        executionId: number;
+        returnCode: number;
+    }
+
+    export class StreamInformation {
+        getAllProperties(): Record<string, string>;
+    }
+
+    export class MediaInformation {
+        getStreams(): StreamInformation[];
+
+        getMediaProperties(): Record<string, string>;
+
+        getAllProperties(): Record<string, string>;
     }
 
     export class RNFFmpegConfig {
-        static getFFmpegVersion(): Promise<{version: string}>;
-        static getPlatform(): Promise<{platform: string}>;
+        static getFFmpegVersion(): Promise<string>;
+
+        static getPlatform(): Promise<string>;
+
         static disableRedirection(): void;
-        static getLogLevel(): Promise<{level: number}>;
+
+        static getLogLevel(): Promise<number>;
+
         static setLogLevel(level: number): void;
+
         static disableLogs(): void;
+
         static disableStatistics(): void;
-        static enableLogCallback(newCallback: (log: LogMessage) => void): void;
+
+        static enableLogCallback(newCallback: (log: Log) => void): void;
+
         static enableStatisticsCallback(newCallback: (statistics: Statistics) => void): void;
+
         static getLastReceivedStatistics(): Promise<Statistics>;
+
         static resetStatistics(): void;
+
         static setFontconfigConfigurationPath(path: string): void;
+
         static setFontDirectory(path: string, mapping?: { [key: string]: string }): void;
-        static getPackageName(): Promise<{packageName: string}>;
+
+        static getPackageName(): Promise<string>;
+
         static getExternalLibraries(): Promise<string[]>;
-        static getLastReturnCode(): Promise<{lastRc: number}>;
-        static getLastCommandOutput(): Promise<{lastCommandOutput: string}>;
-        static registerNewFFmpegPipe(): Promise<{packageName: string}>;
+
+        static getLastReturnCode(): Promise<number>;
+
+        static getLastCommandOutput(): Promise<string>;
+
+        static registerNewFFmpegPipe(): Promise<string>;
+
+        static setEnvironmentVariable(name: string, value: string);
+
+        static writeToPipe(inputPath: string, pipePath: string): Promise<number>;
     }
 
     export class RNFFmpeg {
-        static executeWithArguments(arguments: string[]): Promise<{rc: number}>;
-        static execute(command: string): Promise<{rc: number}>;
+        static executeWithArguments(arguments: string[]): Promise<number>;
+
+        static execute(command: string): Promise<number>;
+
+        static executeAsyncWithArguments(arguments: string[], callback: (execution: CompletedExecution) => void): Promise<number>;
+
+        static executeAsync(command: string, callback: (execution: CompletedExecution) => void): Promise<number>;
+
         static cancel(): void;
+
+        static cancelExecution(executionId: number): void;
+
+        static listExecutions(): Promise<Execution[]>;
+
         static parseArguments(command: string): string[];
     }
 
     export class RNFFprobe {
-        static executeWithArguments(arguments: string[]): Promise<{rc: number}>;
-        static execute(command: string): Promise<{rc: number}>;
+        static executeWithArguments(arguments: string[]): Promise<number>;
+
+        static execute(command: string): Promise<number>;
+
         static getMediaInformation(path: string): Promise<MediaInformation>;
     }