+++ CHANGELOG.md
@@ -1,6 +1,17 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
+## 0.5.0 - 2020-10-22
+- Depends on mobile-ffmpeg v4.4
+- Implements async FFmpeg execution methods
+- Allows modifying mobile-ffmpeg version for android
+- Adds listExecutions API method
+- Updates getMediaInformation implementation
+- Adds setEnvironmentVariable API method
+- LogCallback and StatisticsCallback functions updated with executionId
+- Includes an updated example application
+- Fixes issue #156, #164 and #188
+
 ## 0.4.4 - 2020-03-09
 - Fixes issue #118